requirements.txt
# NanoClaw 의존성 패키지
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
pdfplumber>=0.10.0
tabula-py>=2.8.0
camelot-py>=0.11.0
openpyxl>=3.1.0
lxml>=4.9.0
slack-bolt>=1.18.0
slack-sdk>=3.27.0
python-dotenv>=1.0.0
